Update struct task_attr_str for 200D (D678?)
Not yet tested, this updates the struct based on rom analysis.
This is used in is_taskid_valid().  Usage is the same on 200D and R.

Struct has additional fields, some look important.  Need to audit
how ML uses this struct, since it is not aware of the change, we
may leave fields uninitialised.